remove and refactor contracts from options to config
