+++ packages/contracts/src/http/cookie-config-builder.ts
@@ -1,5 +1,5 @@
 
-import { CookieOptions } from './cookie-options.js'
+import { CookieConfig } from './cookie-config.js'
 
 export type RequestCookieBuilderCallback = (cookieBuilder: RequestCookieBuilder) => void
 export type ResponseCookieBuilderCallback = (cookieBuilder: ResponseCookieBuilder) => void
@@ -71,5 +71,5 @@ export interface ResponseCookieBuilder {
   /**
    * Merge the given `config` with the default HTTP cookie config.
    */
-  useConfig (config: Partial<CookieOptions>): this
+  useConfig (config: Partial<CookieConfig>): this
 }
