+++ packages/contracts/src/http/cookie-config.ts
@@ -1,5 +1,5 @@
 
-export interface CookieOptions {
+export interface CookieConfig {
   /**
    * The time or date in the future at which the cookie expires.
    */
