+++ packages/contracts/src/index.ts
@@ -20,19 +20,19 @@ export { EnvStore } from './env/env.js'
 export { Bootstrapper, BootstrapperCtor } from './core/bootstrapper.js'
 export { ErrorHandler, ErrorHandlerCtor } from './core/error-handler.js'
 
-export { HashBuilder, HashBuilderCallback, HashBuilderOptions } from './hashing/hash-builder.js'
+export { HashBuilder, HashBuilderCallback, HashBuilderConfig } from './hashing/hash-builder.js'
 export { HashConfig } from './hashing/config.js'
 export { BaseHasher } from './hashing/base-hasher.js'
 export { Hasher } from './hashing/hasher.js'
 
-export { BodyparserConfig, BodyparserOptions } from './http/bodyparser-config.js'
+export { BodyparserConfig } from './http/bodyparser-config.js'
 export { HttpConfig } from './http/config.js'
 export { HttpContext, NextHandler } from './http/context.js'
 export { HttpController } from './http/controller.js'
 export { CookieBag } from './http/cookie-bag.js'
-export { CookieOptions } from './http/cookie-options.js'
-export { RequestCookieBuilder, RequestCookieBuilderCallback, ResponseCookieBuilder, ResponseCookieBuilderCallback } from './http/cookie-options-builder.js'
-export { CorsConfig, CorsOptions } from './http/cors-config.js'
+export { CookieConfig } from './http/cookie-config.js'
+export { RequestCookieBuilder, RequestCookieBuilderCallback, ResponseCookieBuilder, ResponseCookieBuilderCallback } from './http/cookie-config-builder.js'
+export { CorsConfig } from './http/cors-config.js'
 export { InteractsWithContentTypes } from './http/concerns/interacts-with-content-types.js'
 export { InteractsWithState } from './http/concerns/interacts-with-state.js'
 export { StateBag, HttpStateData } from './http/concerns/state-bag.js'
@@ -52,7 +52,7 @@ export { HttpRouteGroup } from './http/route-group.js'
 export { HttpRoute, RouteObjectAttributes } from './http/route.js'
 export { HttpRouter, RouteHandler, RouteAttributes } from './http/router.js'
 export { HttpServer, HttpServerHandler } from './http/server.js'
-export { StaticAssetsConfig, StaticAssetsOptions } from './http/static-assets-config.js'
+export { StaticAssetsConfig } from './http/static-assets-config.js'
 export { UploadedFile } from './http/uploaded-file.js'
 
 export { LoggingConfig, LogChannelConfig, LoggingChannels, ConsoleChannelConfig, FileChannelConfig } from './logging/config.js'
